Refactor SSL/SL distinction to dataset level.
Let datasets (ex. `MultiPhysicsDataset`), and thus DataLoaders, handle the
SSL/SL distinction (i.e. whether we are learning to reconstruct the same input
or predict the corresponding output, which is the next trajectory). After this
refactor, the trainer methods don't need to handle this. However, the
wrapper/adjoint model arch. still needs to know which version of `forward()` to
call, so it can't be stage-agnostic.